Refactored concurrent modification exception

+++ product-service/src/main/java/com/example/productservice/entity/Product.java
@@ -19,7 +19,7 @@ public class Product {
     @Column(name = "created_by")
     private String createdBy;
     @Version
-    private Long version;
+    private int version;
 
 }
